Delegate delete to parent to perform any checks

+++ code/extensions/ElementPublishChildren.php
@@ -14,12 +14,19 @@ class ElementPublishChildren extends DataExtension {
 			$widget->publish('Stage', 'Live');
 		}
 
-		// remove any elements that are on live but not in draft.
+		// remove any elements that are on live but not in draft or have been
+		// unlinked from everything
 		$widgets = Versioned::get_by_stage('BaseElement', 'Live', "ParentID = '". $this->owner->ID ."' OR ListID = '".$this->owner->ID."'");
 
 		foreach($widgets as $widget) {
 			if(!in_array($widget->ID, $staged)) {
-				$widget->deleteFromStage('Live');
+				if($this->owner->hasMethod('shouldCleanupElement')) {
+					if($this->owner->shouldCleanupElement($widget)) {
+						$widget->deleteFromStage('Live');
+					}
+				} else {
+					$widget->deleteFromStage('Live');
+				}
 			}
 		}
 	}
